# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LinkedYou Backend API
  description: LinkedYou 后端当前已实现 REST API 的 OpenAPI 3.0 文档。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: 查询用户列表
      description: 查询所有未软删除用户，当前接口未分页。
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
              examples:
                success:
                  summary: 用户列表响应
                  value:
                    success: true
                    code: OK
                    message: success
                    data:
                      - id: 1
                        userId: admin
                        displayName: 系统管理员
                        role: Admin
                        email: admin@example.com
                        phone: "13900000001"
                        avatar: /avatars/admin.png
                        active: true
                        lastLoginAt: "2026-06-26T18:30:00"
                        lastLoginIp: 127.0.0.1
                        createdAt: "2026-06-01T09:00:00"
                        updatedAt: "2026-06-26T18:30:00"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Users
      operationId: createUser
      summary: 创建用户
      description: 创建新用户。`userId` 在未软删除用户中必须唯一，密码不会在响应中返回。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
            examples:
              createUser:
                summary: 创建用户请求
                value:
                  userId: postman_user_001
                  password: PostmanPass123!
                  displayName: Postman Test User
                  role: Admin
                  email: postman_user_001@example.com
                  phone: "13900001001"
                  avatar: /avatars/postman.png
                  active: true
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseEnvelope"
        "400":
          description: 参数校验失败或用户ID重复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  summary: 参数校验失败
                  value:
                    success: false
                    code: VALIDATION_ERROR
                    message: "userId: must not be blank; password: must not be blank"
                    data: null
                userIdExists:
                  summary: 用户ID重复
                  value:
                    success: false
                    code: USERNAME_EXISTS
                    message: username already exists
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/login:
    post:
      tags:
        - Users
      operationId: loginUser
      summary: 用户登录
      description: 校验 `userId` 和密码。成功后写入 `user_sessions`，返回 access token、refresh token 和用户信息；成功、失败、禁用和锁定都会写入 `user_login_logs`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              login:
                summary: 登录请求
                value:
                  userId: admin
                  password: Password123!
                  deviceId: postman-device
                  deviceName: Postman Runtime
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseEnvelope"
              examples:
                success:
                  summary: 登录成功响应
                  value:
                    success: true
                    code: OK
                    message: success
                    data:
                      accessToken: eyJleGFtcGxl
                      refreshToken: refresh_example
                      tokenType: Bearer
                      issuedAt: "2026-06-27T23:30:00"
                      expiresAt: "2026-06-28T00:00:00"
                      refreshExpiresAt: "2026-07-27T23:30:00"
                      user:
                        id: 1
                        userId: admin
                        displayName: 系统管理员
                        role: Admin
                        email: admin@example.com
                        active: true
        "400":
          description: 参数校验失败、用户ID或密码错误、账号禁用或账号锁定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                loginFailed:
                  summary: 用户ID或密码错误
                  value:
                    success: false
                    code: LOGIN_FAILED
                    message: invalid user id or password
                    data: null
                accountDisabled:
                  summary: 账号禁用
                  value:
                    success: false
                    code: ACCOUNT_DISABLED
                    message: account is disabled
                    data: null
                accountLocked:
                  summary: 账号锁定
                  value:
                    success: false
                    code: ACCOUNT_LOCKED
                    message: account is locked
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/refresh-token:
    post:
      tags:
        - Users
      operationId: refreshUserToken
      summary: 刷新 access token
      description: 使用有效 refresh token 刷新 access token。当前实现不轮换 refresh token，数据库中的 `refresh_token_hash` 保持不变，响应会返回原 refresh token。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            examples:
              refresh:
                summary: 刷新 token 请求
                value:
                  refreshToken: refresh_example
      responses:
        "200":
          description: access token 刷新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseEnvelope"
              examples:
                success:
                  summary: 刷新成功响应
                  value:
                    success: true
                    code: OK
                    message: success
                    data:
                      accessToken: new_access_token_example
                      refreshToken: refresh_example
                      tokenType: Bearer
                      issuedAt: "2026-06-28T11:30:00"
                      expiresAt: "2026-06-28T12:00:00"
                      refreshExpiresAt: "2026-07-27T23:30:00"
                      user:
                        id: 1
                        userId: admin
                        displayName: 系统管理员
                        role: Admin
                        email: admin@example.com
                        active: true
        "400":
          description: 参数校验失败、refresh token 无效、用户不存在或账号禁用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tokenInvalid:
                  summary: token 无效
                  value:
                    success: false
                    code: TOKEN_INVALID
                    message: token is invalid
                    data: null
                accountDisabled:
                  summary: 账号禁用
                  value:
                    success: false
                    code: ACCOUNT_DISABLED
                    message: account is disabled
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/logout:
    post:
      tags:
        - Users
      operationId: logoutUser
      summary: 用户登出
      description: 根据 `Authorization` Bearer token 撤销 `user_sessions` 中的当前会话。
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer access token。
          schema:
            type: string
          example: Bearer eyJleGFtcGxl
      responses:
        "200":
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoidResponse"
        "400":
          description: 缺少、格式错误、已过期或已撤销的 token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                tokenInvalid:
                  summary: token 无效
                  value:
                    success: false
                    code: TOKEN_INVALID
                    message: token is invalid
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}:
    get:
      tags:
        - Users
      operationId: getUserById
      summary: 查询单个用户
      description: 根据主键 ID 查询未软删除用户。
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseEnvelope"
        "400":
          description: ID 为负数、用户不存在或用户已软删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                userNotFound:
                  summary: 用户不存在
                  value:
                    success: false
                    code: USER_NOT_FOUND
                    message: user not found
                    data: null
                validationError:
                  summary: ID 参数校验失败
                  value:
                    success: false
                    code: VALIDATION_ERROR
                    message: "getById.id: must be greater than or equal to 0"
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - Users
      operationId: updateUser
      summary: 更新用户
      description: 更新指定用户。请求体字段均为可选字段，只更新传入字段。
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
            examples:
              updateUser:
                summary: 更新用户请求
                value:
                  userId: postman_user_001_updated
                  password: PostmanPass456!
                  displayName: Postman Updated User
                  role: Editor
                  email: updated_postman_user_001@example.com
                  phone: "13900001001"
                  avatar: /avatars/postman-updated.png
                  active: false
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseEnvelope"
        "400":
          description: 参数校验失败、用户不存在或用户ID重复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                userNotFound:
                  summary: 用户不存在
                  value:
                    success: false
                    code: USER_NOT_FOUND
                    message: user not found
                    data: null
                validationError:
                  summary: 参数校验失败
                  value:
                    success: false
                    code: VALIDATION_ERROR
                    message: "password: size must be between 8 and 100"
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Users
      operationId: deleteUser
      summary: 软删除用户
      description: 软删除指定用户，写入 `deleted_at`，不会物理删除数据库记录。
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoidResponse"
              examples:
                success:
                  summary: 删除成功响应
                  value:
                    success: true
                    code: OK
                    message: success
                    data: null
        "400":
          description: ID 为负数、用户不存在或用户已软删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                userNotFound:
                  summary: 用户不存在
                  value:
                    success: false
                    code: USER_NOT_FOUND
                    message: user not found
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
  /api/users/{id}/password:
    put:
      tags:
        - Users
      operationId: changeUserPassword
      summary: 更改用户密码
      description: 校验当前密码后更改用户密码。新密码按 `application.yml` 中 `security.password-policy` 规则校验，规则可通过 `enabled` 开关启用或关闭。
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserChangePasswordRequest"
            examples:
              changePassword:
                summary: 更改密码请求
                value:
                  currentPassword: PostmanPass456!
                  newPassword: PostmanPass789!
      responses:
        "200":
          description: 更改成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoidResponse"
              examples:
                success:
                  summary: 更改成功响应
                  value:
                    success: true
                    code: OK
                    message: success
                    data: null
        "400":
          description: 参数校验失败、用户不存在、当前密码错误或新密码不符合策略
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                currentPasswordInvalid:
                  summary: 当前密码错误
                  value:
                    success: false
                    code: CURRENT_PASSWORD_INVALID
                    message: current password is invalid
                    data: null
                passwordPolicyViolation:
                  summary: 新密码不符合策略
                  value:
                    success: false
                    code: PASSWORD_POLICY_VIOLATION
                    message: "password does not match policy: Password must contain at least 1 special characters."
                    data: null
                userNotFound:
                  summary: 用户不存在
                  value:
                    success: false
                    code: USER_NOT_FOUND
                    message: user not found
                    data: null
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    UserIdPath:
      name: id
      in: path
      required: true
      description: 用户主键 ID，必须大于等于 0。
      schema:
        type: integer
        format: int64
        minimum: 0
      example: 1
...
  responses:
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalError:
              summary: 内部错误响应
              value:
                success: false
                code: INTERNAL_ERROR
                message: internal server error
                data: null
  schemas:
    LoginRequest:
      type: object
      required:
        - userId
        - password
      properties:
        userId:
          type: string
          description: 登录用户ID。
          maxLength: 64
          example: admin
        password:
          type: string
          description: 登录密码。
          maxLength: 100
          format: password
          example: Password123!
        deviceId:
          type: string
          nullable: true
          description: 前端设备 ID。
          maxLength: 128
          example: postman-device
        deviceName:
          type: string
          nullable: true
          description: 前端设备名称。
          maxLength: 150
          example: Postman Runtime
        ipAddress:
          type: string
          nullable: true
          description: 登录 IP。前端不传时后端从请求中获取。
          maxLength: 45
          example: 127.0.0.1
        userAgent:
          type: string
          nullable: true
          description: User-Agent。前端不传时后端从请求头获取。
          maxLength: 500
          example: PostmanRuntime/7
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: 登录或上次刷新返回的 refresh token。当前实现不轮换该 token。
          maxLength: 512
          example: refresh_example
    UserCreateRequest:
      type: object
      required:
        - userId
        - password
      properties:
        userId:
          type: string
          description: 登录用户ID，未软删除用户中必须唯一。
          maxLength: 64
          example: postman_user_001
        password:
          type: string
          description: 登录密码，服务端使用 BCrypt 哈希后保存，响应中不会返回。
          minLength: 8
          maxLength: 100
          format: password
          example: PostmanPass123!
        displayName:
          type: string
          nullable: true
          description: 显示名称或昵称。
          maxLength: 100
          example: Postman Test User
        role:
          type: string
          nullable: true
          description: 用户角色。
          maxLength: 100
          example: Admin
        email:
          type: string
          nullable: true
          description: 邮箱地址。
          format: email
          maxLength: 191
          example: postman_user_001@example.com
        phone:
          type: string
          nullable: true
          description: 手机号。
          maxLength: 32
          example: "13900001001"
        avatar:
          type: string
          nullable: true
          description: 头像路径。
          maxLength: 255
          example: /avatars/postman.png
        active:
          type: boolean
          nullable: true
          description: 是否启用。不传时默认为 `true`。
          example: true
    UserUpdateRequest:
      type: object
      description: 更新用户请求。所有字段均可选，只更新传入字段。
      properties:
        userId:
          type: string
          nullable: true
          description: 登录用户ID，未软删除用户中必须唯一。
          maxLength: 64
          example: postman_user_001_updated
        password:
          type: string
          nullable: true
          description: 新登录密码，服务端使用 BCrypt 哈希后保存，响应中不会返回。
          minLength: 8
          maxLength: 100
          format: password
          example: PostmanPass456!
        displayName:
          type: string
          nullable: true
          description: 显示名称或昵称。
          maxLength: 100
          example: Postman Updated User
        role:
          type: string
          nullable: true
          description: 用户角色。
          maxLength: 100
          example: Editor
        email:
          type: string
          nullable: true
          description: 邮箱地址。
          format: email
          maxLength: 191
          example: updated_postman_user_001@example.com
        phone:
          type: string
          nullable: true
          description: 手机号。
          maxLength: 32
          example: "13900001001"
        avatar:
          type: string
          nullable: true
          description: 头像路径。
          maxLength: 255
          example: /avatars/postman-updated.png
        active:
          type: boolean
          nullable: true
          description: 是否启用。
          example: false
    UserChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: 当前密码。
          maxLength: 100
          format: password
          example: PostmanPass456!
        newPassword:
          type: string
          description: 新密码，按 `security.password-policy` 配置校验。
          maxLength: 100
          format: password
          example: PostmanPass789!
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: 访问 token。当前为不透明随机 token。
          example: eyJleGFtcGxl
        refreshToken:
          type: string
          description: 刷新 token。数据库只保存其 SHA-256 摘要。
          example: refresh_example
        tokenType:
          type: string
          example: Bearer
        issuedAt:
          type: string
          format: date-time
          example: "2026-06-27T23:30:00"
        expiresAt:
          type: string
          format: date-time
          description: access token 过期时间，由 `security.auth.access-token-ttl` 控制。
          example: "2026-06-28T00:00:00"
        refreshExpiresAt:
          type: string
          format: date-time
          description: refresh token 过期时间，由 `security.auth.refresh-token-ttl` 控制。
          example: "2026-07-27T23:30:00"
        user:
          $ref: "#/components/schemas/UserResponse"
...
    UserResponse:
      type: object
      description: 用户响应数据，不包含密码哈希、软删除时间等内部字段。
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: string
          example: admin
        displayName:
          type: string
          nullable: true
          example: 系统管理员
        role:
          type: string
          nullable: true
          example: Admin
        email:
          type: string
          nullable: true
          format: email
          example: admin@example.com
        phone:
          type: string
          nullable: true
          example: "13900000001"
        avatar:
          type: string
          nullable: true
          example: /avatars/admin.png
        active:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          nullable: true
          format: date-time
          example: "2026-06-26T18:30:00"
        lastLoginIp:
          type: string
          nullable: true
          example: 127.0.0.1
        createdAt:
          type: string
          nullable: true
          format: date-time
          example: "2026-06-01T09:00:00"
        updatedAt:
          type: string
          nullable: true
          format: date-time
          example: "2026-06-26T18:30:00"
...
    UserResponseEnvelope:
      type: object
      required:
        - success
        - code
        - message
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: "#/components/schemas/UserResponse"
    LoginResponseEnvelope:
      type: object
      required:
        - success
        - code
        - message
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: "#/components/schemas/LoginResponse"
    UserListResponse:
      type: object
      required:
        - success
        - code
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
...
    VoidResponse:
      type: object
      required:
        - success
        - code
        - message
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          nullable: true
          example: null
    ErrorResponse:
      type: object
      required:
        - success
        - code
        - message
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          description: 错误码，例如 `VALIDATION_ERROR`、`USER_NOT_FOUND`、`USERNAME_EXISTS`、`INTERNAL_ERROR`。
          example: USER_NOT_FOUND
        message:
          type: string
          description: 错误消息。
          example: user not found
        data:
          nullable: true
          example: null